Separate integration test file into separate modules
